Create Pair with given Sum (Two Sum)
The Two Sum Problem is a classic programming challenge that involves finding two numbers in a given list (array) that add up to a specific target value. Instead of just identifying the numbers, the task is to return the indices of these numbers within the list. Or in the case of this problem, return True if there is a pair of elements in the array.